# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Bauk Cheescake",
        "version": "0.0.11",
        "description": "appiid: 3232f21e-653e-4e98-b46d-c34cbc751182 - TODO: All users can post new addresses"
    },
    "servers": [
        {
            "url": "https://localhost:443"
        }
    ],
    "security": [
        {
            "BearerAuth": [],
            "Refresh": [],
            "MailAuth": [],
            "AppiId": []
        }
    ],
    "paths": {
...
        "/users/places": {
            "get": {
                "security": [
                    {
                        "AppiId": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Get addresses",
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "Application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetPlaces"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/places/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": [],
                        "AppiId": []
                    }
                ],
                "description": "Get all places",
                "tags": [
                    "Places"
                ],
                "responses": {
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "Errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "msg": {
                                                        "type": "string",
                                                        "example": "Invalid value"
                                                    },
                                                    "param": {
                                                        "type": "string",
                                                        "example": "AreaCode"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "body"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Places"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": [],
                        "AppiId": []
                    }
                ],
                "requestBody":{
                    "content": {
                        "Application/json":{
                            "schema":{
                                "$ref":"#/components/schemas/PlacesPost"
                            }
                        }
                    }
                },
                "description": "Add new place",
                "tags": [
                    "Places"
                ],
                "responses": {
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "Errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "msg": {
                                                        "type": "string",
                                                        "example": "Invalid value"
                                                    },
                                                    "param": {
                                                        "type": "string",
                                                        "example": "authorization"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "headers"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetUser"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/places/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": [],
                        "AppiId": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "description": "Get place by id",
                "tags": [
                    "Places"
                ],
                "responses": {
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "Errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "msg": {
                                                        "type": "string",
                                                        "example": "Invalid value"
                                                    },
                                                    "param": {
                                                        "type": "string",
                                                        "example": "authorization"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "headers"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Places"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/pictures/": {
            "post": {
                "security": [
                    {
                        "AppiId": [],
                        "BearerAuth": []
                    }
                ],
                "description": "",
                "requestBody": {
                    "content": {
                        "Application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Pictures"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Pictures"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorizaiton"
                    }
                }
            },
            "get": {
                "security": [
                    {
                        "AppiId": []
                    }
                ],
                "description": "Get picture of product",
                "tags": [
                    "Pictures"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }            
        },
        "/pictures/{id}": {
            "get": {
                "security": [
                    {
                        "AppiId": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "description": "Get picture by product id",
                "tags": [
                    "Pictures"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }            
        },
...
        "/orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": [],
                        "AppiId": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "description": "Get addition",
                "tags": [
                    "Orders"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": [],
                        "AppiId": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "Application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Addition"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "description": "Get addition",
                "tags": [
                    "Orders"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/orders/": {
            "post": {
                "security": [
                    {
                        "AppiId": [],
                        "BearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "Application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Orders"
                            }
                        }
                    }
                },
                "description": "Post orders",
                "tags": [
                    "Orders"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "get": {
                "security": [
                    {
                        "AppiId": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Get order",
                "tags": [
                    "Orders"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "Application/json":{
                                "schema":{
                                    "$ref": "#/components/schemas/Orders"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "Refresh": {
                "type": "apiKey",
                "in": "header",
                "name": "Refresh"
            },
            "MailAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "MailToken"
            },
            "AppiId": {
                "type": "apiKey",
                "in": "header",
                "name": "appiId"
            }
        },
        "schemas": {
...
            "GetUser": {
                "type": "object",
                "required": [
                    "ID",
                    "FirstName",
                    "LastName",
                    "PhoneNumber",
                    "Email",
                    "IdStatus",
                    "IdRole"
                ],
                "properties": {
                    "ID": {
                        "type": "string"
                    },
                    "FirstName": {
                        "type": "string"
                    },
                    "LastName": {
                        "type": "string"
                    },
                    "PhoneNumber": {
                        "type": "string"
                    },
                    "Email": {
                        "type": "string"
                    },
                    "Status": {
                        "type": "object",
                        "required": [
                            "ID",
                            "Status"
                        ],
                        "properties": {
                            "ID": {
                                "type": "integer"
                            },
                            "Status": {
                                "type": "string"
                            }
                        }
                    },
                    "Role": {
                        "type": "object",
                        "required": [
                            "ID",
                            "RoleName"
                        ],
                        "properties": {
                            "ID": {
                                "type": "integer"
                            },
                            "RoleName": {
                                "type": "string"
                            }
                        }
                    },
                    "Addresses": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "required": [
                                "ID",
                                "StreetAndNumber",
                                "GpsLocation",
                                "PlaceName",
                                "AreaCode",
                                "Active"
                            ],
                            "properties": {
                                "ID": {
                                    "type": "integer"
                                },
                                "StreetAndNumber": {
                                    "type": "string"
                                },
                                "GpsLocation": {
                                    "type": "string"
                                },
                                "PlaceName": {
                                    "type": "string"
                                },
                                "AreaCode": {
                                    "type": "string"
                                },
                                "Active": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            },
...
            "GetPlaces": {
                "type": "array",
                "items": {
                    "type": "object",
                    "required": [
                        "ID",
                        "PlaceName",
                        "AreaCode",
                        "Active"
                    ],
                    "properties": {
                        "ID": {
                            "type": "integer"
                        },
                        "PlaceName": {
                            "type": "string"
                        },
                        "AreaCode": {
                            "type": "string"
                        },
                        "Active": {
                            "type": "boolean"
                        }
                    }
                }
            },
...
            "Addition": {
                "type": "object",
                "properties": {
                    "Name": {
                        "type": "string"
                    },
                    "Price": {
                        "type": "integer"
                    }
                }
            },
            "Orders": {
                "type": "object",
                "properties": {
                    "Description": {
                        "type": "string"
                    },
                    "IdVisit": {
                        "type": "integer"
                    },
                    "IdAddress": {
                        "type": "integer"
                    },
                    "IdOrderStatus":{
                        "type": "integer"
                    },
                    "Price": {
                        "type": "integer"
                    },
                    "Items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Item"
                        }
                    }
                }
            },
            "Item": {
                "type": "object",
                "properties": {
                    "IdProduct": {
                        "type": "integer"
                    },
                    "IdSize": {
                        "type": "integer"
                    },
                    "Price": {
                        "type": "integer"
                    },
                    "Additions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AdditionGet"
                        }
                    }
                }
            },
            "AdditionGet": {
                "type": "object",
                "properties": {
                    "ID": {
                        "type": "integer"
                    },
                    "Name": {
                        "type": "string"
                    },
                    "Price": {
                        "type": "integer"
                    }
                }
            },
            "Places": {
                "type": "object",
                "properties": {
                    "ID": {
                        "type": "integer"
                    },
                    "PlaceName": {
                        "type": "string"
                    },
                    "AreaCode": {
                        "type": "integer"
                    },
                    "Active": {
                        "type": "boolean"
                    }
                }
            },
            "PlacesPost": {
                "type": "object",
                "properties": {
                    "PlaceName": {
                        "type": "string"
                    },
                    "AreaCode": {
                        "type": "integer"
                    },
                    "Active": {
                        "type": "boolean"
                    }
                }
            },
...
            "Pictures":{
                "type": "object",
                "properties": {
                    "Name": {
                        "type": "string"
                    },
                    "File_Path": {
                        "type": "string"
                    }
                }
            }
        }
    }
}